break-out dedicated Logout component
